fix modules folders, started tests usecases
