+++ src/@clean/modules/notebook/domain/repositories/notebook_repository_interface.ts
@@ -1,0 +1,5 @@
+import { Notebook } from "../../../../shared/domain/entities/notebook";
+
+export interface INotebookRepository {
+  getNotebook(numSerie: string): Promise<Notebook>;
+}
