--- src/@clean/modules/user/domain/repositories/notebook_repository_interface.ts
@@ -1,5 +1,0 @@
-import { Notebook } from "@/@clean/shared/domain/entities/notebook";
-
-export interface INotebookRepository {
-    getNotebook(numSerie: string): Promise<Notebook>;
-}
