--- src/@clean/modules/user/usecases/get_all_withdraws_usecase.ts
@@ -1,14 +1,0 @@
-import { NoItemsFoundError } from "@/@clean/shared/domain/helpers/errors/domain_error";
-import { IWithdrawRepository } from "../domain/repositories/withdraw_repository_interface";
-
-export class GetAllWithdraws {
-    constructor(private withdrawRepo: IWithdrawRepository) {}
-
-    async execute() {
-        const withdraws = await this.withdrawRepo.getAllWithdraws();
-        if (withdraws.length == 0) {
-            throw Error('Withdraws empty')
-        }
-        return withdraws;
-    }
-}
