+++ src/@clean/modules/withdraw/usecases/get_all_withdraws_usecase.ts
@@ -1,0 +1,13 @@
+import { IWithdrawRepository } from "../domain/repositories/withdraw_repository_interface";
+
+export class GetAllWithdraws {
+  constructor(private withdrawRepo: IWithdrawRepository) {}
+    
+  async execute() {
+    const withdraws = await this.withdrawRepo.getAllWithdraws();
+    if (withdraws.length == 0) {
+      throw Error('Withdraws empty')
+    }
+    return withdraws;
+  }
+}
